// src/shared/consts/react-query.d.ts
/**
 * query Key
 */
export declare const queryKey: {
    infinite: string;
    list: string;
    user: string;
    product: string;
    quantity: string;
    category: string;
    category_list: string;
    image: string;
    cart: string;
    order: string;
    order_detail: string;
    pay_history: string;
};
export declare const staleTime: {
    product: number;
    cart: number;
    order_detail: number;
};